Lagt til endepunkt og laget ny service for å håndtere utvidet behandling.
